fix: Move AspectQueryParser to domain layer (Clean Architecture)
Resolves layering violation by moving query parsing logic from application layer to domain service layer.

Changes:
- Move AspectQueryParser from application/query to domain/service/query
- Move AspectQueryAST and ComparisonOperator to domain/service/query
- Move QueryParseError to domain/service/query
- Update all import statements in dependent files

Closes #177
